begin update to visual studio 2015 and .net 4
project is now in .net 4.  Repaired the unit tests (they are now VS 2015
style unit tests with C# annotations) and they now run against a
provided .net 4 assembly. (work still in progress).  currently the
compile feature which appears to be mono based is not working in the
unit tests and is commented out. my intention is to try to port randoop
to a modern open source C# roslyn and .net core tool.